### 2020-01-26 #2041 @whedon
[PRE REVIEW]: hypr: An R package for hypothesis-driven contrast coding
**Submitting author:** @mmrabe (<a href="http://orcid.org/0000-0002-2556-5644">Maximilian Rabe</a>)
**Repository:** <a href="https://github.com/mmrabe/hypr" target ="_blank">https://github.com/mmrabe/hypr</a>
**Version:** v0.1.6
