### 2020-01-26 #2041 @labarba
👋 @marcosvital — Could you handle this JOSS submission as editor? Let me know!
