### 2020-02-06 #2041 @marcosvital
Hi, @mmrabe, I'm Marcos Vital, and I'll be the editor of your submission.

Did you had the chance to check the article proof generated above? If everything seems to be ok with it, we can move on to assigning reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this.

You can also check [this list of people](https://bit.ly/joss-reviewers) that have agreed to review for JOSS, and see if any of them may be suitable to review this submission.

And let me know if you need any help finding reviewers if you don't have anyone in mind.
