### 2020-02-07 7ffccc3
Update paper bibliography
