+++ paper.bib
@@ -1,6 +1,6 @@
 @Article{Schad2020,
   author    = {Daniel J. Schad and Shravan Vasishth and Sven Hohenstein and Reinhold Kliegl},
-  title     = {How to capitalize on a priori contrasts in linear (mixed) models: A tutorial},
+  title     = {How to capitalize on a priori contrasts in linear (mixed) models: {A} tutorial},
   journal   = {Journal of Memory and Language},
   year      = {2020},
   volume    = {110},
@@ -10,7 +10,7 @@
 @Book{Venables2002,
   author    = {W. N. Venables and B. D. Ripley},
   year      = {2002},
-  title     = {Modern Applied Statistics with S},
-  edition   = {Fourth Edition},
+  title     = {Modern Applied Statistics with {S}},
+  edition   = {Fourth ed},
   publisher = {Springer},
 }
