### 2020-02-07 #2041 @mmrabe
Hi @marcosvital,

Thanks a lot for handling this submission. The paper proof looked fine except for two minor issues in the bibliography, which I just fixed.

I have browsed the list of potential reviewers and would like to suggest the following (in no particular order):

- tomfaulkenberry
- MikeLydeamore
- njtierney
- rmhogervorst
- JonathanReardon
- lindemann09
- janfreyberg

Best regards,
Max
