### 2020-02-14 4ed40d4
+, * and ** methods
