+++ NAMESPACE
@@ -16,7 +16,9 @@ export(hmat2cmat)
 export(hmat2eqs)
 export(hypr)
 export(thmat)
+exportMethods("*")
 exportMethods("+")
+exportMethods("^")
 exportMethods("formula<-")
 exportMethods("levels<-")
 exportMethods("names<-")
