+++ DESCRIPTION
@@ -23,7 +23,7 @@ Suggests:
 VignetteBuilder: knitr
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 6.1.1
+RoxygenNote: 7.0.2
 Collate: 
     'equations.R'
     'hypr.R'
