### 2020-02-15 2e71500
Add / method, swap * and ** so that * is compatible with standard R syntax
