+++ NAMESPACE
@@ -18,6 +18,7 @@ export(hypr)
 export(thmat)
 exportMethods("*")
 exportMethods("+")
+exportMethods("/")
 exportMethods("^")
 exportMethods("formula<-")
 exportMethods("levels<-")
