### 2020-02-16 6688767
Use & instead of ** or ^ for interaction
