+++ man/combination.Rd
@@ -3,7 +3,7 @@
 \name{+,hypr,hypr-method}
 \alias{+,hypr,hypr-method}
 \alias{*,hypr,hypr-method}
-\alias{^,hypr,hypr-method}
+\alias{&,hypr,hypr-method}
 \alias{/,hypr,hypr-method}
 \title{Combining hypr objects by addition or interaction}
 \usage{
@@ -11,7 +11,7 @@
 
 \S4method{*}{hypr,hypr}(e1, e2)
 
-\S4method{^}{hypr,hypr}(e1, e2)
+\S4method{&}{hypr,hypr}(e1, e2)
 
 \S4method{/}{hypr,hypr}(e1, e2)
 }
@@ -28,7 +28,7 @@ You can combine one or more \code{hypr} objects, i.e. combine their hypothesis t
 \itemize{
 \item \code{*,hypr,hypr-method}: Interaction of \code{e1} and \code{e2}
 
-\item \code{^,hypr,hypr-method}: Interaction and main contrasts of \code{e1} and \code{e2}
+\item \code{&,hypr,hypr-method}: Interaction and main contrasts of \code{e1} and \code{e2}
 
 \item \code{/,hypr,hypr-method}: Nesting levels of \code{e2} within \code{e1}
 }}
@@ -43,8 +43,8 @@ hc <- h1 + h2
 
 hc
 
-interaction <- h1 * h2
+interaction <- h1 & h2
 
-interaction_and_main <- h1 ** h2
+interaction_and_main <- h1 * h2
 
 }
