+++ NAMESPACE
@@ -16,10 +16,10 @@ export(hmat2cmat)
 export(hmat2eqs)
 export(hypr)
 export(thmat)
+exportMethods("&")
 exportMethods("*")
 exportMethods("+")
 exportMethods("/")
-exportMethods("^")
 exportMethods("formula<-")
 exportMethods("levels<-")
 exportMethods("names<-")
