### 2020-02-28 #2041 @marcosvital
Dear @rmhogervorst and @JonathanReardon

Would you be willing to review this submission for JOSS?
If you are still new to this, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is still a pre-review, and once there are enough reviewers, we will move on to a new issue where the review will take place.
