### 2020-02-28 c75c8a8
Fix bug in vignette
