+++ vignettes/hypr-regression.Rmd
@@ -35,12 +35,13 @@ For the examples in this vignette, we are using a simulated dataset with one fac
 
 ```{r}
 set.seed(123)
-M <- c(X1 = 10, X2 = 20, X3 = 10, X4 = 40) # condition means
+M <- c(mu1 = 10, mu2 = 20, mu3 = 10, mu4 = 40) # condition means
 N <- 5
 SD <- 10
 simdat <- do.call(rbind, lapply(names(M), function(x) {
   data.frame(X = x, DV = as.numeric(MASS::mvrnorm(N, unname(M[x]), SD^2, empirical = TRUE)))
 }))
+simdat$X <- factor(simdat$X)
 simdat$id <- 1:nrow(simdat)
 simdat
 ```
