### 2020-03-02 #2134 @marcosvital
Dear @tomfaulkenberry and @JonathanReardon, thank you very much for accepting review this submission for JOSS.

Please check the instructions and checklists above, and let me know if you need any assistance. You can also tag @mmrabe  if you need to ask specific questions about the submission.
