### 2020-03-09 #2134 @JonathanReardon
Hello, I can't seem to access the checklist at all (unable to tick any boxes), could you look into this for me? apologies if it's a fault my end.
