### 2020-03-10 #2134 @marcosvital
Hi, @JonathanReardon, no need to apologize!

In order to be able to tick the boxes, you need to accept the invite as a collaborator on the joss-reviews repository - GitHub should have sent you an invite. This is easily missed depending on how you handle notifications from GitHub, so I'll send you another one.

Let me know if this works, ok?
