### 2020-03-29 #2134 @tomfaulkenberry
I have (finally) completed my review of this paper/package.  It is a very nice package, and will be immediately useful as a tool for psychologists who want to translate the (sometimes unfamiliar) R method of specifying contrasts to the way that we are usually taught to specify contrasts.

I left two boxes unchecked in my review above: (1) there is no license in the Github repo, and (2) there is no statement of how people can contribute.  These should be easy to fix.

Functionally, the package installed easily (even on my half-broken installation of R), and the functions did exactly what they were supposed to.

Overall, nice package, and good paper.

Tom
