### 2020-03-30 #2134 @JonathanReardon
I too have now finished the review, apologies for the delay, and thank you to all for being understanding. This was a very simple tool to install and use, it worked exactly as intended and behaved exactly as described. No issues at all.

If you could include a license and a brief statement as to how others can contribute to the project, that would be great. I (and @tomfaulkenberry) could then tick off the remaining 2 boxes.

Nice paper and useful tool that I am sure many psychologists would be interested in using.

Thanks all,
Jonathan
