### 2020-04-02 15cb870
Add community guidelines to README
