### 2020-04-02 38ff80d
Add community guidelines to Rmd README
