### 2020-04-02 #2134 @mmrabe
@tomfaulkenberry and @JonathanReardon, thank you both very much for taking the time to review the package and paper and of course for your wonderful feedback! @marcosvital, I hope it’s okay that I went ahead and made the changes to the repository.

I added a “Community guidelines” paragraph to the end of the [README](https://github.com/mmrabe/hypr/blob/master/README.md) file and a [LICENSE](https://github.com/mmrabe/hypr/blob/master/LICENSE) file (GPL-3) to the root directory. Some of these details can also be found in the [DESCRIPTION](https://github.com/mmrabe/hypr/blob/master/DESCRIPTION) meta file.

This was a great experience so far and I really enjoyed the interaction. I’m looking forward to hearing from you again.

Thanks a lot and stay safe!
Max
