### 2020-04-09 #2134 @marcosvital
@mmrabe, everything is ok with cited DOIs.

I generated a new proof, so you can take a carefull (maybe final?) look at the manuscript. I'll read it myself too, and will let you know if I find any issues, ok?
