### 2020-04-24 #2134 @marcosvital
Hi, @mmrabe! Sorry about the late reply. Let's move on, everything looks fine with the manuscript.

As one of our last steps, if you already didn't do this, you will need to archive the last release of the package (on Zenodo, figshare, or other of your choice). After that, post the version number and archive DOI here, and we can continue.
