+++ DESCRIPTION
@@ -3,7 +3,7 @@ Type: Package
 Title: Hypothesis Matrix Translation
 URL: https://maxrabe.com/hypr
 BugReports: https://github.com/mmrabe/hypr/issues
-Version: 0.1.6
+Version: 0.1.7
 Authors@R: c(
     person(given = "Maximilian M.", family = "Rabe", email = "maximilian.rabe@uni-potsdam.de", role = c("aut","cre")),
     person(given = "Shravan", family = "Vasishth", role = c("aut")),
@@ -23,7 +23,7 @@ Suggests:
 VignetteBuilder: knitr
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.0.2
+RoxygenNote: 7.1.0
 Collate: 
     'equations.R'
     'hypr.R'
