### 2020-04-24 #2134 @mmrabe
Hi @marcosvital! Thanks! I just archived the package using Zenodo. The current version number is v0.1.7 (was v0.1.6 upon submission) and the Zenodo DOI is: 10.5281/zenodo.3765843
