### 2020-04-27 #2134 @danielskatz
👋 @mmrabe - please update the metadata in the Zenodo archive, specifically the title, to match the JOSS paper
