### 2020-04-27 3bb538c
Add package paper DOI
